refactor: reorganize DTO imports into a dedicated request directory for better structure
